refactor: extract projects routes from create_app closure
Move the user-facing /api/me + /api/projects routes and the admin
project CRUD/members/CI-bot family out of the create_app closure and
into routes/projects.py and routes/admin_projects.py, following the
existing routes/ extraction pattern. Route table is unchanged in
content and order (161 routes before and after).
